# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Pedido:
    properties:
      cliente_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      observacao:
        type: string
      produtos:
        items:
          $ref: '#/definitions/domain.Produto'
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Produto:
    properties:
      categoria:
        type: string
      created_at:
        type: string
      descricao:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.StatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a documentation of all endpoints in the API.
  title: Tech Challenge API
  version: 3.0.0
paths:
...
  /pedido/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: id do pedido
        in: path
        name: id
        required: true
        type: integer
      - description: 'status permitido: recebido, em_preparacao, pronto, finalizado'
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/domain.StatusRequest'
      produces:
      - application/json
      responses: {}
      summary: atualiza o status do pedido
      tags:
      - Pedido
...
  /pedido/detail/{id}:
    get:
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id do pedido a ser lista
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Pedido'
      summary: lista detalhes do pedido
      tags:
      - Pedido
...
  /pedidos/{statuses}:
    get:
      parameters:
      - description: status dos pedidos a ser pesquisado:(recebido, em_preparacao,
          pronto, finalizado)
        in: path
        name: statuses
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Pedido'
            type: array
      summary: lista pedido por status
      tags:
      - Pedido
...
produces:
- application/json
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: token
    type: apiKey
swagger: "2.0"